it is now workable. Naked Singles, Bifurcation, and one optional technique: hidden singles are implemented. Naked singles are mandatory since hidden singles rely on them to do the crosshatching, though this may be changed later

+++ src/Index.hpp
@@ -1,6 +1,6 @@
 #ifndef INDEX_HPP
 #define INDEX_HPP
 
-typedef unsigned Index_t;
+typedef unsigned short Index_t;
 
 #endif
